requirements.txt
accelerate==1.11.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
appdirs==1.4.4
asttokens==3.0.0
attrs==25.4.0
backoff==2.2.1
bcrypt==5.0.0
bitsandbytes==0.48.2
build==1.3.0
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
cfgv==3.4.0
charset-normalizer==3.4.4
chromadb==1.1.1
click==8.3.0
colorama==0.4.6
coloredlogs==15.0.1
cryptography==46.0.3
decorator==5.2.1
diskcache==5.6.3
distlib==0.4.0
distro==1.9.0
docstring_parser==0.17.0
durationpy==0.10
et_xmlfile==2.0.0
executing==2.2.1
fastuuid==0.14.0
filelock==3.20.0
flatbuffers==25.9.23
frozenlist==1.8.0
fsspec==2025.9.0
google-auth==2.42.0
googleapis-common-protos==1.71.0
grpcio==1.76.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.36.0
humanfriendly==10.0
identify==2.6.15
idna==3.11
importlib_metadata==8.7.0
importlib_resources==6.5.2
instructor==1.12.0
ipython==9.6.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
json5==0.12.1
json_repair==0.25.2
jsonpatch==1.33
jsonpickle==4.1.1
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kubernetes==34.1.0
langchain==1.0.2
langchain-core==1.0.1
langgraph==1.0.2
langgraph-checkpoint==3.0.0
langgraph-prebuilt==1.0.2
langgraph-sdk==0.2.9
langsmith==0.4.38
litellm==1.79.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
mcp==1.19.0
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
multidict==6.7.0
networkx==3.5
nodeenv==1.9.1
numpy==2.3.4
oauthlib==3.3.1
onnxruntime==1.23.2
openai==1.109.1
openpyxl==3.1.5
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-grpc==1.38.0
opentelemetry-exporter-otlp-proto-http==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
orjson==3.11.4
ormsgpack==1.11.0
overrides==7.7.0
packaging==25.0
parso==0.8.5
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==12.0.0
platformdirs==4.5.0
portalocker==2.7.0
posthog==5.4.0
pre_commit==4.3.0
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.0
psutil==7.1.2
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pycparser==2.23
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
Pygments==2.19.2
PyJWT==2.10.1
pypdfium2==5.0.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
pyvis==0.3.2
pywin32==311
PyYAML==6.0.3
referencing==0.37.0
regex==2025.10.23
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.2.0
rpds-py==0.28.0
rsa==4.9.1
safetensors==0.6.2
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sse-starlette==3.0.2
stack-data==0.6.3
starlette==0.49.1
sympy==1.14.0
tenacity==9.1.2
tiktoken==0.12.0
tokenizers==0.22.1
tomli==2.3.0
tomli_w==1.2.0
torch==2.9.0
torchvision==0.24.0
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.1
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.3.0
uv==0.9.5
uvicorn==0.38.0
virtualenv==20.35.4
watchfiles==1.1.1
wcwidth==0.2.14
websocket-client==1.9.0
websockets==15.0.1
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
